+++ integration/read-file/tests/test_transforms.py
@@ -42,17 +42,34 @@ def klio_msg():
 @pytest.fixture
 def expected_log_messages(klio_msg):
     return [
-        (
-            "KlioThreadLimiter(name=LogKlioMessage.process) Blocked – "
-            "waiting on semaphore for an available thread (available threads:"
-        ),
-        (
-            "KlioThreadLimiter(name=LogKlioMessage.process) Released "
-            "semaphore (available threads:"
-        ),
-        "Hello, Klio!",
-        "Received element {}".format(klio_msg.data.element),
-        "Received payload {}".format(klio_msg.data.payload),
+        {
+            "level": "DEBUG",
+            "message": (
+                "KlioThreadLimiter(name=LogKlioMessage.process) Blocked – "
+                "waiting on semaphore for an available thread (available threads:"
+            ),
+        },
+        {
+            "level": "DEBUG",
+            "message": (
+                "KlioThreadLimiter(name=LogKlioMessage.process) Released "
+                "semaphore (available threads:"
+            ),
+        },
+        {
+            "level": "DEBUG",
+            "message": "Loading config file from "
+            "/usr/local/klio-job-run-effective.yaml.",
+        },
+        {"level": "INFO", "message": "Hello, Klio!"},
+        {
+            "level": "INFO",
+            "message": "Received element {}".format(klio_msg.data.element),
+        },
+        {
+            "level": "INFO",
+            "message": "Received payload {}".format(klio_msg.data.payload),
+        },
     ]
 
 
@@ -64,6 +81,6 @@ def test_process(klio_msg, expected_log_messages, caplog):
     assert len(caplog.records) == len(expected_log_messages)
 
     for index, record in enumerate(caplog.records):
-        assert "INFO" == record.levelname
-        assert expected_log_messages[index] in record.message
-
+        expected_log_message = expected_log_messages[index]
+        assert expected_log_message["level"] == record.levelname
+        assert expected_log_message["message"] in record.message