+++ docs/src/reference/core/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+21.2.0 (2021-03-16)
+-------------------
+
+Changed
+*******
+
+* Moved ``IndentListDumper`` to ``klio_core`` config utils.
+
+
 0.2.2 (2021-02-11)
 ------------------
 
@@ -10,6 +19,7 @@ Fixed
 * Fixed handling of configured BQ schemas when defined as a string (thank you `gfalcone <https://github.com/spotify/klio/pull/165>`_!).
 * Fixed config bug that skipped preprocessing (overrides, templates) of dict parsed from YAML
 * ``KlioWriteToAvro`` has been enabled as an output event type (previously missing).
+
 
 0.2.1 (2020-12-03)
 ------------------