+++ docs/src/reference/executor/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+21.2.0 (2021-03-16)
+-------------------
+
+Fixed
+*****
+
+* ``klioexec`` now writes runtime config to include in ``setup.py`` distribution.
+
+
 0.2.2 (2021-01-14)
 ------------------